Adding foreign keys Changing key types from bigint to int
git-tfs-id: [https://tfs.codeplex.com/tfs/tfs04]$/polaris;C38263

+++ Main/Polaris.Dbl/Schema Objects/Tables/Polaris.PlayLog.table.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [Polaris].[PlayLog]
 (
-[PlayLogId] [bigint] NOT NULL IDENTITY(1, 1),
-[GameId] [bigint] NOT NULL,
-[UserId] [bigint] NOT NULL,
+[PlayLogId] [int] NOT NULL IDENTITY(1, 1),
+[GameId] [int] NOT NULL,
+[UserId] [int] NOT NULL,
 [Date] [datetime] NOT NULL,
 [Score] [decimal] (18, 0) NOT NULL
 ) ON [PRIMARY]
